Create ViewModel for HappyBirthdayScreen.kt
